Remove Pod-chaser book-marks...
For whatever reason, both Spotify and Pod-chaser's watch-later play-lists fail to load when empty. 

Who the fuck is coding these?! My best guess is they are using reactive-frameworks with for-like reactive-rendering and completely failing to handle the empty case in planning; coding and testing.

Great stuff! )Spotify is known to use React, by the way).